requirements.txt
datasets>=2.14.0
scikit-learn>=1.3.0
PyYAML>=6.0
PyMuPDF==1.26.6
beautifulsoup4==4.14.2
langchain==1.0.5
langchain-community==0.4.1
langchain-text-splitters
sentence-transformers
qdrant-client
google-generativeai==0.8.5
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=1.0.0
langchain-huggingface>=0.0.1
langchain-text-splitters>=0.0.1
qdrant-client>=1.7.0
langchain-qdrant>=0.1.0
sentence-transformers>=2.2.0
google-generativeai>=0.3.0
google-search-results>=2.4.2
transformers>=4.35.0
python-dotenv>=1.0.0
streamlit>=1.30.0
torch>=2.0.0